corregido el boton de eliminacion de registros

+++ resources/views/cursos/show.blade.php
@@ -42,11 +42,11 @@
 
 
                             </dl>
-
-                            {{ Form::open(['route'=>['cursos.destroy',$curso],'method'=>'delete']) }}
-                            {{ Form::submit('Eliminar Curso',['class'=>'btn btn-danger']) }}
-                            {{ Form::close() }}
-
+                            @if(Auth::check() and Auth::user()->role == 'admin')
+                                {{ Form::open(['route'=>['cursos.destroy',$curso],'method'=>'delete']) }}
+                                {{ Form::submit('Eliminar Curso',['class'=>'btn btn-danger']) }}
+                                {{ Form::close() }}
+                            @endif
                             <hr>
                         </div>
